SeansValidator yazıldı.

+++ EntityLayer/Bilet.cs
@@ -17,7 +17,7 @@ namespace EntityLayer
         public string seyirciSoyad { get; set; }
         [StringLength(11)]
         public string seyirciTc { get; set; }
-        public int fiyat { get; set; }
+        public double fiyat { get; set; }
         [StringLength(50)]
         public string seriNo { get; set; }
         public bool odemeTipi { get; set; }
